Support parsing multiple lines (#42)
Fixes #40

Call the parser until either an error occurs, or the s-expression is
parsed successfully. I think the downside here is that a list parsing
error in the message from the solver could put the parser into a state
where it's parsing forever, but I'm not sure that we can really detect
that case without any other framing.