Add support for parsing negative solution values (#46)
## Description

This PR would add support for solutions that contain negative integers. 

Currently, the problem with negative solutions is that the SMT solver
will return for example `(x (- 2))` a solution to `x < 0` . When calling
`get`, this solution will be parsed into an `SExprData::List` . Calling
`try_from_int` will then return always return the error `NotAnAtom`.

This PR introduces a new variant of `SExprData`, which is a special case
of `SExprData::List`, which is returned for lists that contain exactly
two atoms , where the first one is either `+` or `-`.

Additionally it fixes typos in comments.


## Minimal Test

```rust
use easy_smt::{ContextBuilder, Response};

#[test]
fn test_negative_integers() {
    let mut ctx = ContextBuilder::new()
        .solver("z3")
        .solver_args(["-smt2", "-in"])
        .build()
        .unwrap();

    let x = ctx.declare_const("x", ctx.int_sort()).unwrap();

    ctx.assert(ctx.lt(x, ctx.numeral(0))).unwrap();

    assert_eq!(ctx.check().unwrap(), Response::Sat);

    let solution = ctx.get_value(vec![x]).unwrap();
    let sol_x = ctx.get(solution[0].1);

    let sol = i64::try_from(sol_x).expect("Failed to parse");
    assert!(sol < 0)
}
```

---------

Co-authored-by: Nick Fitzgerald <fitzgen@gmail.com>